Old digi module split into digi and readout.
