+++ docs/digi.rst
@@ -1,20 +1,16 @@
 :mod:`hexsample.digi` --- Digitization facilities
 =================================================
 
-This module contains all the digitization facilities, including the event structure
-for a digitized events (:class:`DigiEvent <hexsample.digi.DigiEvent>`) and a generic
-event readout structure (:class:`HexagonalReadout <hexsample.digi.HexagonalReadout>`).
+This module contains...
 
 Digi event structure
 --------------------
 
 .. literalinclude:: ../hexsample/digi.py
-   :pyobject: DigiEvent
+   :pyobject: DigiEventRectangular
    :end-before: __post_init__
 
 
-Digitization process
---------------------
